+++ .gitignore
@@ -43,6 +43,8 @@ captures/
 .idea/caches
 .idea/misc.xml
 .idea/modules.xml
+.idea/copyright/HADA.xml
+.idea/copyright/profiles_settings.xml
 
 # Keystore files
 # Uncomment the following line if you do not want to check your keystore files in.
