+++ .idea/copyright/ProjectCopyright.xml
@@ -1,0 +1,6 @@
+<component name="CopyrightManager">
+  <copyright>
+    <option name="myName" value="ProjectCopyright" />
+    <option name="notice" value="CMPUT 301 - Fall 2018&#10;&#10;Version 1.0&#10;&#10;&amp;#36;today&#10;&#10;This is a group project for CMPUT 301 course at the University of Alberta&#10;Copyright (C) &amp;#36;today.year  Austin Goebel, Anders Johnson, Alex Li,&#10;Cristopher Penner, Joseph Potentier-Neal, Jason Robock&#10;Last modified &amp;#36;file.lastModified" />
+  </copyright>
+</component>
