+++ app/src/main/AndroidManifest.xml
@@ -2,8 +2,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="ca.ualberta.cs.cmput301f18t19.hada.hada">
 
-    <uses-permission android:name="android.permission.INTERNET" />
-
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -18,16 +16,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.NewUserActivity" />
-        <activity android:name=".ui.ViewPatientProblemsActivity" />
         <activity android:name=".ui.ViewProblemActivity" />
+        <activity android:name=".ui.NewUserActivity" />
         <activity android:name=".ui.PatientListActivity" />
         <activity android:name=".ui.AddRecordActivity" />
-        <activity android:name=".ui.AddPatientActivity" />
-        <activity android:name=".ui.ViewRecordActivity" />
-        <activity android:name=".ui.EditUserSettingsActivity" />
-        <activity android:name=".ui.ProblemListActivity"/>
-        <activity android:name=".ui.AddProblemActivity" />
     </application>
-    <uses-permission android:name="android.permission.INTERNET" />
+
 </manifest>
