+++ app/src/main/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/CareProvider.java
@@ -22,13 +22,6 @@ public class CareProvider extends User {
 
     private ArrayList<Patient> patients = new ArrayList<Patient>();
 
-    public CareProvider(){
-        super();
-    }
-    public CareProvider(String userID, String phoneNumber, String emailAddress){
-        super(userID, phoneNumber, emailAddress);
-    }
-
     public void setPatients(ArrayList<Patient> patients){
         this.patients = patients;
     }
