+++ app/src/main/java/ca/ualberta/cs/cmput301f18t19/hada/hada/ui/MainActivity.java
@@ -13,24 +13,28 @@
 
 package ca.ualberta.cs.cmput301f18t19.hada.hada.ui;
 
-import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
+import android.text.Editable;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.Toast;
 
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
+
+import java.io.BufferedReader;
+import java.io.FileInputStream;
+import java.io.InputStreamReader;
+import java.lang.reflect.Type;
 import java.util.ArrayList;
-import java.util.concurrent.ExecutionException;
 
 import ca.ualberta.cs.cmput301f18t19.hada.hada.R;
 import ca.ualberta.cs.cmput301f18t19.hada.hada.model.CareProvider;
-import ca.ualberta.cs.cmput301f18t19.hada.hada.model.ESUserManager;
-import ca.ualberta.cs.cmput301f18t19.hada.hada.model.LoggedInSingleton;
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.ListManagerPatient;
 import ca.ualberta.cs.cmput301f18t19.hada.hada.model.Patient;
-import ca.ualberta.cs.cmput301f18t19.hada.hada.model.UserController;
+
+import static android.provider.Telephony.Mms.Part.FILENAME;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -38,9 +42,6 @@ public class MainActivity extends AppCompatActivity {
     private EditText usernameInfo;
     private ArrayList<Patient> patientList = new ArrayList<Patient>();
     private ArrayList<CareProvider> careProviderList = new ArrayList<CareProvider>();
-    private Patient patient;
-    private CareProvider careProvider;
-
 
 
 
@@ -49,28 +50,19 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        usernameInfo = (EditText) findViewById(R.id.mainActivityUsernameText);
-        final Button patientLogin = findViewById(R.id.mainActivityPatientLogin);
-        Button careProviderLogin = findViewById(R.id.mainActivityDoctorLogin);
-        Button createUser = findViewById(R.id.mainActivityCreateUser);
-
+        usernameInfo = (EditText) findViewById(R.id.username_login);
+        Button patientLogin = findViewById(R.id.patient_login);
+        Button careProviderLogin = findViewById(R.id.doctor_login);
+        Button createUser = findViewById(R.id.create_user);
 
 
         patientLogin.setOnClickListener(new View.OnClickListener() {
             public void onClick(View v) {
 
                 String username = usernameInfo.getText().toString();
-                Patient patient = new UserController().getPatient(username);
-
-                if(patient != null){
-                        Log.d("Username logged in", patient.getUserID());
-                        LoggedInSingleton.getInstance().setLoggedInID(patient.getUserID());
-                        Intent intent = new Intent(MainActivity.this, ProblemListActivity.class);
-
-                        startActivity(intent);
-                }
-                else{Toast.makeText(MainActivity.this, "Username does not exist. Create a new user instead!?", Toast.LENGTH_SHORT).show();}
-
+                /**
+                 * Pull patient from username patient ArrayList
+                 */
             }
         });
 
@@ -78,33 +70,11 @@ public class MainActivity extends AppCompatActivity {
             public void onClick(View v) {
 
                 String username = usernameInfo.getText().toString();
-                ESUserManager.GetCareProviderTask careProviderTask = new ESUserManager.GetCareProviderTask();
-                careProviderTask.execute(username);
-
-                try {
-                    CareProvider careProvider = careProviderTask.get();
-
-                    if(careProvider != null){
-                        Log.d("Username logged in", careProvider.getUserID());
-                        LoggedInSingleton.getInstance().setLoggedInID(careProvider.getUserID());
-                        Intent intent = new Intent(MainActivity.this, PatientListActivity.class);
-                        startActivity(intent);
-                    }
-                    else{Toast.makeText(MainActivity.this, "Username does not exist. Create a new user instead!?", Toast.LENGTH_SHORT).show();}
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                } catch (ExecutionException e) {
-                    e.printStackTrace();
-                }
+                /**
+                 * Pull doctor from username patient ArrayList
+                 */
             }
 
-        });
-
-        createUser.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v){
-                Intent intent = new Intent(MainActivity.this, NewUserActivity.class);
-                startActivity(intent);
-            }
         });
 
     }
