+++ app/src/main/java/ca/ualberta/cs/cmput301f18t19/hada/hada/ui/ViewProblemActivity.java
@@ -15,6 +15,7 @@ package ca.ualberta.cs.cmput301f18t19.hada.hada.ui;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
 import ca.ualberta.cs.cmput301f18t19.hada.hada.R;
 
 public class ViewProblemActivity extends AppCompatActivity {
