--- app/src/main/res/layout/activity_add_problem.xml
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="ui.AddProblemActivity">
-
-    <TextView
-        android:id="@+id/addProblemActivityTitle"
-        android:layout_width="334dp"
-        android:layout_height="32dp"
-        android:layout_marginTop="16dp"
-        android:gravity="center_horizontal"
-        android:text="@string/add_problem_title"
-        android:textAppearance="@style/TextAppearance.AppCompat.Caption"
-        android:textSize="24sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <EditText
-        android:id="@+id/addProblemTitle"
-        android:layout_width="match_parent"
-        android:layout_height="40dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="24dp"
-        android:layout_marginEnd="16dp"
-        android:hint="@string/add_problem_hint_title"
-        android:importantForAutofill="no"
-        android:inputType="text"
-        android:singleLine="true"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/addProblemActivityTitle" />
-
-    <EditText
-        android:id="@+id/addProblemDate"
-        android:layout_width="match_parent"
-        android:layout_height="40dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="10dp"
-        android:layout_marginEnd="16dp"
-        android:hint="@string/add_problem_hint_date"
-        android:importantForAutofill="no"
-        android:inputType="date"
-        android:singleLine="true"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.514"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/addProblemTitle" />
-
-    <EditText
-        android:id="@+id/addProblemDescription"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="10dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="10dp"
-        android:layout_marginEnd="16dp"
-        android:background="@null"
-        android:gravity="top|start"
-        android:hint="@string/add_problem_hint_description"
-        android:importantForAutofill="no"
-        android:inputType="textMultiLine"
-        android:lines="10"
-        android:maxWidth="200dp"
-        android:maxLines="15"
-        android:minLines="5"
-        android:scrollHorizontally="false"
-        android:singleLine="true"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/addProblemDate" />
-
-    <Button
-        android:id="@+id/addProblemButton"
-        android:layout_width="250dp"
-        android:layout_height="40dp"
-        android:layout_marginBottom="40dp"
-        android:text="@string/add_problem_title"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</android.support.constraint.ConstraintLayout>
