+++ app/src/main/res/layout/activity_main.xml
@@ -7,59 +7,59 @@
     tools:context=".ui.MainActivity">
 
     <Button
-        android:id="@+id/mainActivityPatientLogin"
+        android:id="@+id/patient_login"
         android:layout_width="150dp"
         android:layout_height="wrap_content"
+        android:layout_marginTop="24dp"
         android:text="@string/patient_login"
-        app:layout_constraintBottom_toTopOf="@+id/mainActivityDoctorLogin"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintHorizontal_bias="0.509"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/mainActivityUsernameText" />
+        app:layout_constraintTop_toBottomOf="@+id/username_login" />
 
     <Button
-        android:id="@+id/mainActivityDoctorLogin"
+        android:id="@+id/doctor_login"
         android:layout_width="150dp"
         android:layout_height="wrap_content"
+        android:layout_marginTop="8dp"
+        android:layout_marginBottom="16dp"
         android:text="@string/doctor_login"
-        app:layout_constraintBottom_toTopOf="@+id/mainActivityCreateUser"
+        app:layout_constraintBottom_toTopOf="@+id/create_user"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintHorizontal_bias="0.501"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/mainActivityPatientLogin" />
+        app:layout_constraintTop_toBottomOf="@+id/patient_login"
+        app:layout_constraintVertical_bias="1.0" />
 
     <Button
-        android:id="@+id/mainActivityCreateUser"
+        android:id="@+id/create_user"
         android:layout_width="150dp"
-        android:layout_height="48dp"
-        android:layout_marginBottom="32dp"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="36dp"
         android:text="@string/create_user"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/mainActivityDoctorLogin" />
+        app:layout_constraintStart_toStartOf="parent" />
 
     <EditText
-        android:id="@+id/mainActivityUsernameText"
+        android:id="@+id/username_login"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="40dp"
         android:ems="10"
-        android:hint="@string/enter_username"
         android:inputType="textPersonName"
-        app:layout_constraintBottom_toTopOf="@+id/mainActivityPatientLogin"
+        android:text="@string/enter_username"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintHorizontal_bias="0.502"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/mainActivityImage" />
+        app:layout_constraintTop_toBottomOf="@+id/imageView" />
 
     <ImageView
-        android:id="@+id/mainActivityImage"
+        android:id="@+id/imageView"
         android:layout_width="150dp"
         android:layout_height="141dp"
-        android:layout_marginTop="32dp"
-        app:layout_constraintBottom_toTopOf="@+id/mainActivityUsernameText"
+        android:layout_marginTop="48dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@mipmap/ic_launcher_foreground" />
