+++ app/src/main/res/layout/activity_view_problem.xml
@@ -49,17 +49,14 @@
 
     <ListView
         android:id="@+id/viewProblemList"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
+        android:layout_width="368dp"
+        android:layout_height="184dp"
         android:layout_marginStart="8dp"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="8dp"
         android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
-        android:background="@color/colorListBackground"
         android:divider="@color/colorPrimaryDark"
         android:headerDividersEnabled="false"
         android:scrollbars="horizontal"
-        app:layout_constraintBottom_toTopOf="@+id/ViewProblemSlideshowButton"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/viewProblemImage" />
@@ -81,14 +78,15 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
         android:layout_marginEnd="8dp"
         android:layout_marginBottom="8dp"
         android:backgroundTint="@color/colorAccent"
         android:contentDescription="@string/ViewProblemSlideshowButton"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toStartOf="@+id/viewProblemAddRecordButton"
-        app:layout_constraintHorizontal_bias="0.481"
         app:layout_constraintStart_toEndOf="@+id/viewProblemMapButton"
+        app:layout_constraintTop_toBottomOf="@+id/viewProblemList"
         app:srcCompat="@drawable/ic_slideshow_white_24dp" />
 
 </android.support.constraint.ConstraintLayout>
