+++ app/src/main/res/values/colors.xml
@@ -4,5 +4,4 @@
     <color name="colorPrimaryDark">#00574b</color>
     <color name="colorAccent">#d81b60</color>
     <color name="colorListBackground">#ececec</color>
-    <color name="Disabled_Field">#cccccc</color>
 </resources>
