--- app/src/test/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/ESPatientManagerTest.java
@@ -1,5 +1,0 @@
-package ca.ualberta.cs.cmput301f18t19.hada.hada.model;
-
-public class ESPatientManagerTest {
-
-}
