+++ app/src/test/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/ListManagerPatientTest.java
@@ -10,17 +10,19 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.concurrent.ExecutionException;
 
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.ListManagerPatient;
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.Patient;
 import io.searchbox.client.JestResult;
 import io.searchbox.core.Search;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
-
-public class ESProblemManagerTest {
+public class ListManagerPatientTest {
     @Test
-    public void testSaveProblem() {
-        Problem problem = new Problem();
-        new ESProblemManager.AddProblemsTask().execute(problem);
+    public void testSavePatient() {
+        Patient patient = new Patient();
+        new ListManagerPatient.AddPatientTask().execute(patient);
 
         DroidClientConfig config = new DroidClientConfig
                 .Builder("http://cmput301.softwareprocess.es:8080/cmput301f18t19test")
@@ -44,19 +46,19 @@ public class ESProblemManagerTest {
     }
 
     @Test
-    public void testGetProblem() {
-        Problem problem = new Problem();
-        ArrayList<Problem> problems = new ArrayList<Problem>();
-        new ESProblemManager.AddProblemsTask().execute(problem);
+    public void testGetPatient() {
+        Patient patient = new Patient();
+        ArrayList<Patient> patients = new ArrayList<Patient>();
+        new ListManagerPatient.AddPatientTask().execute(patient);
 
         try {
-            problems = new ESProblemManager.GetProblemsTask().execute("").get();
+            patients = new ListManagerPatient.GetPatientsTask().execute("").get();
         } catch (ExecutionException e) {
             e.printStackTrace();
         } catch (InterruptedException e) {
             e.printStackTrace();
         }
-        assertFalse(problems.isEmpty());
+        assertFalse(patients.isEmpty());
 
     }
 }
