+++ app/src/test/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/ListManagerRecordsTest.java
@@ -12,6 +12,8 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.concurrent.ExecutionException;
 
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.ListManagerRecord;
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.Record;
 import io.searchbox.client.JestResult;
 import io.searchbox.core.Search;
 
@@ -20,11 +22,11 @@ import static org.junit.Assert.*;
 //http://cmput301.softwareprocess.es:8080/cmput301f18t19
 //http://cmput301.softwareprocess.es:8080/cmput301f18t19test
 
-public class ESRecordManagerTest {
+public class ListManagerRecordsTest {
     @Test
     public void testSaveRecord() {
         Record record = new Record();
-        new ESRecordManager.AddRecordTask().execute(record);
+        new ListManagerRecord.AddRecordTask().execute(record);
         //so we gotta check if it is in the db so we should search for it
         DroidClientConfig config= new DroidClientConfig
                 //testing for our group
@@ -47,9 +49,9 @@ public class ESRecordManagerTest {
     public void testGetRecordTask(){
         Record record = new Record();
         ArrayList<Record> records = new ArrayList<Record>();
-        new ESRecordManager.AddRecordTask().execute(record);
+        new ListManagerRecord.AddRecordTask().execute(record);
         try {
-            records = new ESRecordManager.GetRecordTask().execute("").get();
+            records = new ListManagerRecord.GetRecordTask().execute("").get();
         } catch (ExecutionException e) {
             e.printStackTrace();
         } catch (InterruptedException e) {
